lld: Make relocation handling compartment-aware
This change starts by splitting a separate SymbolCompartAux object out
of Symbol.  This new structure contains per-compartment fields for
symbols, such as if a symbol needs an entry in a given compartment's
GOT or PLT.  This requires passing compartment pointers to various
Symbol methods which is what most of this commit is.

In theory this should now generate separate GOTs and PLTs for each
compartment.  There are some sharp edges that are probably broken
around copy relocations (at least).