lld: Rework implicit compartment assignment
Build a map to track inter-section dependencies where the entry for
each section is a list of unique <section, relocation type> tuples.
For logging purposes, each of this tuples is mapped to the target
Symbol of the first relocation of the given type.

This map can then be scanned multiple times as compartments are
assigned without having to rebuild the map between assignments.

Add an additional pass of "hard" requirements requiring that input
sections referenced directly by a PCC relocation are assigned to the
same compartment as the PCC reference.

Use a disjoint set union (DSU) to reimplement the existing pass.  Walk
the section dependencies map joining sets for for each relocation
between sections.  However, ignore relocations between two sections
that are already assigned to a compartment since the goal is to
identify sections in the default compartment that are associated with
only a single explicit compartment.  Unlike the previous version, the
sets are built in a single pass and then each set is evaluated as a
single pass.